move createDagEdge to our new EdgeSlice
