+++ src/store/DagNodeSlice/dagNodeUtils.spec.ts
@@ -1,20 +1,9 @@
 import '@testing-library/jest-dom';
-import { Edge } from 'reactflow';
 import { DecisionTree } from 'store/DagNodeSlice/dagNodeSlice';
-import { addDagEdge, applyPositionToNodes, createDagEdge } from 'store/DagNodeSlice/dagNodeUtils';
-import { describe, expect, it, test } from 'vitest';
+import { applyPositionToNodes } from 'store/DagNodeSlice/dagNodeUtils';
+import { describe, expect, it } from 'vitest';
 
-describe('Dag Slice internals', () => {
-  describe('Create Dag Edge', () => {
-    test('takes 2 Id and returns an edge', () => {
-      const sourceId = '2';
-      const targetId = '3';
-      const edge = createDagEdge(sourceId, targetId);
-      expect(typeof edge).toBe('object');
-      expect(edge.source).toBe(sourceId);
-      expect(edge.target).toBe(targetId);
-    });
-  });
+describe('Dag Node Slice internals', () => {
   it('Applies tree positions to nodes', () => {
     const oldX = 0;
     const oldY = 10;
@@ -50,15 +39,5 @@ describe('Dag Slice internals', () => {
     ];
     const nodesWithPositions = applyPositionToNodes(tree, existingNodes);
     expect(nodesWithPositions[0].position).toEqual({ x: newX, y: newY });
-  });
-  it('adding an edge is idempotent', () => {
-    const id1 = '1';
-    const id2 = '2';
-    const id3 = '3';
-    const currentEdges: Edge[] = [{ ...createDagEdge(id1, id2) }, { ...createDagEdge(id2, id3) }];
-    const nonUpdatedEdges = addDagEdge(currentEdges, { source: id1, target: id2 });
-    expect(nonUpdatedEdges).toEqual(currentEdges);
-    const updatedEdges = addDagEdge(currentEdges, { source: id1, target: id3 });
-    expect(updatedEdges.length).toBe(currentEdges.length + 1);
   });
 });
